+++ tests/ci/app/config/bootstrap.php
@@ -19,7 +19,6 @@ use lithium\core\Libraries;
 
 Libraries::add('lithium');
 Libraries::add('app', ['default' => true]);
-Libraries::add('li3_fixtures');
 
 if (file_exists($file = LITHIUM_LIBRARY_PATH . '/autoload.php')) {
 	require_once $file;